Use statemanager in contract (#1896)
* introduce two layer trie

* add unit test

* add unit test

* Using two layer trie in state factory, and refactor working set

* delete NewWorkingSet

* use state/putstate/delstate in contract

+++ db/trie/branchroottrie.go
@@ -141,6 +141,9 @@ func (tr *branchRootTrie) DB() KVStore {
 }
 
 func (tr *branchRootTrie) deleteNodeFromDB(tn Node) error {
+	if tr.root == tn && bytes.Equal(tr.rootHash, tr.emptyRootHash()) {
+		return nil
+	}
 	return tr.kvStore.Delete(tr.nodeHash(tn))
 }
